+++ src/index.js
@@ -2,7 +2,7 @@ module.exports = {
     // Base Class
     Client: require('./Client'),
 
-    Match: require('./Match'),
+    Match: require('./matches/Match'),
     Player: require('./Player'),
 
     version: require('../package.json').version,