+++ src/util/Constants.js
@@ -1,0 +1,18 @@
+module.exports = {
+    BASE_URL: 'https://api.playbattlegrounds.com',
+
+    SHARDS: [
+        'xbox-as',
+        'xbox-eu',
+        'xbox-na',
+        'xbox-oc',
+        'pc-krjp',
+        'pc-na',
+        'pc-eu',
+        'pc-oc',
+        'pc-kakao',
+        'pc-sea',
+        'pc-sa',
+        'pc-as',
+    ],
+};